banner blet qorks for teh nose part

+++ Hooks/HeartPickups.cs
@@ -74,7 +74,7 @@ namespace Avalon.Hooks
             {
                 if (Type == ItemID.Heart && source is not EntitySource_TileBreak && source is not EntitySource_Wiring)
                 {
-                    if (Main.player[p].statLifeMax2 > 400 && Main.player[p].statLifeMax2 < 500 && Main.rand.Next((int)((Main.player[p].statLifeMax2 - 400) / 5), 21) == 20)
+                    if (Main.player[p].statLifeMax2 > 400 && Main.player[p].statLifeMax2 < 500 && Main.rand.Next((Main.player[p].statLifeMax2 - 400) / 5, 21) == 20)
                     {
                         if (Main.player[p].GetModPlayer<AvalonPlayer>().EtherealHeart && Main.rand.NextBool(10) ||
                             Main.player[p].GetModPlayer<AvalonPlayer>().HeartGolem && Main.rand.NextBool(20) ||
@@ -99,7 +99,7 @@ namespace Avalon.Hooks
                 }
                 if (Type == ItemID.CandyApple && source is not EntitySource_TileBreak && source is not EntitySource_Wiring)
                 {
-                    if (Main.player[p].statLifeMax2 > 400 && Main.player[p].statLifeMax2 < 500 && Main.rand.Next((int)((Main.player[p].statLifeMax2 - 400) / 5), 21) == 20)
+                    if (Main.player[p].statLifeMax2 > 400 && Main.player[p].statLifeMax2 < 500 && Main.rand.Next((Main.player[p].statLifeMax2 - 400) / 5, 21) == 20)
                     {
                         if (Main.player[p].GetModPlayer<AvalonPlayer>().EtherealHeart && Main.rand.NextBool(10) ||
                             Main.player[p].GetModPlayer<AvalonPlayer>().HeartGolem && Main.rand.NextBool(20) ||
@@ -124,7 +124,7 @@ namespace Avalon.Hooks
                 }
                 if (Type == ItemID.CandyCane && source is not EntitySource_TileBreak && source is not EntitySource_Wiring)
                 {
-                    if (Main.player[p].statLifeMax2 > 400 && Main.player[p].statLifeMax2 < 500 && Main.rand.Next((int)((Main.player[p].statLifeMax2 - 400) / 5), 21) == 20)
+                    if (Main.player[p].statLifeMax2 > 400 && Main.player[p].statLifeMax2 < 500 && Main.rand.Next((Main.player[p].statLifeMax2 - 400) / 5, 21) == 20)
                     {
                         if (Main.player[p].GetModPlayer<AvalonPlayer>().EtherealHeart && Main.rand.NextBool(10) ||
                             Main.player[p].GetModPlayer<AvalonPlayer>().HeartGolem && Main.rand.NextBool(20) ||
